libモジュールの廃止
- lib.rs削除
- lib.rs内のすべての関数をmain.rsに移行
- Cargo.tomlの[lib]セクションの廃止
