+++ tex/list.tex
@@ -17,10 +17,6 @@
 \inputminted[linenos, breaklines]{rust}{src-rs/main.rs}
 \clearpage
 
-\subsection{ライブラリメインモジュール lib.rs}
-\inputminted[linenos, breaklines]{rust}{src-rs/lib.rs}
-\clearpage
-
 \subsection{アプリケーションステータス app\_status.rs}
 \inputminted[linenos, breaklines]{rust}{src-rs/app_status.rs}
 \clearpage
